core/requirements.txt
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.1.0
matplotlib>=3.8.0
requests>=2.31.0
mysqlclient==2.2.7
openpyxl>=3.1.0
spacy>=3.8.0
cryptography>=41.0.0
